Update patch set 1
Patch Set 1:

> there must be seperate commits somewhere for all this changes

None of the github i've checked did separate commits, the guy is working for Google, code is really well documented. If it's needed that I do a commit for each and every file, i will do, but my comments on those won't be appropriate, i don't have the required C/CPP knowledge to understand the whole thing.

Patch-set: 1